server/requirements.txt
Django == 5.1.1
django-environ == 0.11.2
django-cors-headers == 4.4.0
djangorestframework == 3.15.2
djangorestframework-simplejwt == 5.3.1
django-filter == 24.3
psycopg == 3.2.2
psycopg2-binary == 2.9.9
asgiref== 3.8.1
pip == 24.2
sqlparse == 0.5.1
tzdata == 2024.1
pytest == 8.3.3
